youtube-airflow-etl/requirements.txt
requests
psycopg2-binary
python-dotenv
pytest